Add Robots game by Michel d'Alger
Michel d'Alger <REDACTED>

to me
Adrian:

You'll find attached to this email the ROBOTS game source code from
MSXdev'11 ready to be compiled with asMSX. Consider it to be GPL.

Take care!

- Michel

2013/12/18 Adrian Oboroc <REDACTED>
Hi Michel,

I'd like to ask for the permission to add the source code of your
msxdev'11 game Robots to asmsx project at
https://code.google.com/p/asmsx/

The project is licensed under GPLv3 due to cjv99 choice. cjv99 is the
person who bought all rights to asmsx from pitpan and released is as
free software.

Would you agree to license this file under GPLv3? Please note, this
would be an irrevocable assignment, you won't be able to recall GPLv3
license from it.

I'd like to include this game as example, but if you don't want to, it
is ok, I won't include them.

Thanks,

Adrian Oboroc