Added theNestruo's World Rally code for testing.
MAAAAAAAAAAYBE we should clean all the not required files before
uploading, but I'm at the Revision demoparty and I'm a bit lazy (all I
want is to fix "NOP" bug >:().

So, this is a note for the future me or to someone that wants to
contribute: Clean this! The makefile only build the rom, it doesn't
clean anything.

\o