This apparently solves the NOP bug.
In the `lex.l` when the `IDENTIFICATOR` and `LOCAL_IDENTIFICATOR` are catched
from text the string was directly copied. In fact it only copies the
pointer. If `yytext` is freed it may be overwritten later. Therefore
what I did to fix is to use `strdup` to duplicate the content.

Probably this new string would need to be freed in the future (now it
might be a memory leak). We should investigate this using valgrind.

After some testing I can say this:

Closes #4 [theNestruo] "Add NOPs random" bug