Fused two inclusion rules on parser1
The recognition of include and the actual including of the file was
split in two rules. This leads to the Issue #76.

In order to fix this, I have fused both rules and added that the "\n" at
the end of the line is optional. With this, we can have an include at
the end of the file without an endline.

Closes Issue #76.